third_party/WebKit/LayoutTests/crypto/subtle/rsa-pss/verify-expected.txt

Test verification using RSA-PSS and salt length of 10.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Verifying known signature...
PASS verificationResult is true

Verifying a bad signature...
PASS verificationResult is false
PASS successfullyParsed is true

TEST COMPLETE
